LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/events-004-expected.txt


PASS repeating lists 
FAIL truncating lists assert_equals: expected "padding-top" but got "padding-left"
